shim refactoring
