+++ shims/spark311/src/main/scala/org/apache/spark/sql/rapids/shims/spark311/GpuInMemoryTableScanExec.scala
@@ -17,7 +17,6 @@
 package org.apache.spark.sql.rapids.shims.spark311
 
 import com.nvidia.spark.rapids.GpuExec
-import com.nvidia.spark.rapids.shims.spark311.ParquetCachedBatchSerializer
 
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.catalyst.InternalRow
