+++ shims/spark320/src/main/scala/com/nvidia/spark/rapids/shims/spark320/SparkShimServiceProvider.scala
@@ -14,15 +14,13 @@
  * limitations under the License.
  */
 
-package com.nvidia.spark.rapids.shims.spark311
+package com.nvidia.spark.rapids.shims.spark320
 
 import com.nvidia.spark.rapids.{SparkShims, SparkShimVersion}
 
 object SparkShimServiceProvider {
-  // keep 3.1.0 snapshot version for now until 3.1.1 is released
-  val VERSION310 = SparkShimVersion(3, 1, 0)
-  val VERSION = SparkShimVersion(3, 1, 1)
-  val VERSIONNAMES: Seq[String] = Seq(VERSION310, VERSION)
+  val VERSION320 = SparkShimVersion(3, 2, 0)
+  val VERSIONNAMES: Seq[String] = Seq(VERSION320)
     .flatMap(v => Seq(s"$v", s"$v-SNAPSHOT"))
 }
 
@@ -33,6 +31,6 @@ class SparkShimServiceProvider extends com.nvidia.spark.rapids.SparkShimServiceP
   }
 
   def buildShim: SparkShims = {
-    new Spark311Shims()
+    new Spark320Shims()
   }
 }
